Replace MapQuest w/ tile map support and fix reloading issues
Fixes #46 and #53
