+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 2.5-SNAPSHOT
 
+- Remove MapQuest support
+- Add support for loading tiles from generic tile map servers
+- Fix a bug that would cause tabs to be blank after reloading
 - Add an option to control whether entries should be removed from the DB if they are deleted from the filesystem
 - Missing bags will be deleted by default
 - Updated Spring & Hibernate-related dependencies to newer versions
